move correlationValidation to its own folder
