+++ src/LEGO.AsyncAPI/Validation/Rules/AsyncApiTagRules.cs
@@ -25,24 +25,4 @@ namespace LEGO.AsyncAPI.Validation.Rules
 
                 });
     }
-
-    [AsyncApiRule]
-    public static class AsyncApiCorrelationIdRules
-    {
-        public static ValidationRule<AsyncApiCorrelationId> CorrelationIdRequiredFields =>
-            new ValidationRule<AsyncApiCorrelationId>(
-                (context, correlationId) =>
-                {
-                    context.Enter("location");
-                    if (correlationId.Location == null)
-                    {
-                        context.CreateError(
-                            nameof(CorrelationIdRequiredFields),
-                            string.Format(Resource.Validation_FieldRequired, "location", "correlationId"));
-                    }
-
-                    context.Exit();
-
-                });
-    }
 }
